edb-terraform cli - move script related functions under utils
